adding dashboard#index
