+++ app/controllers/admin/dashboard_controller.rb
@@ -1,0 +1,5 @@
+class Admin::DashboardController < ApplicationController
+
+  def index
+  end
+end
