+++ app/views/admin/dashboard/index.html.erb
@@ -1,0 +1,1 @@
+<h1>Hello There</h1>
