+++ test/controllers/admin/dashboard_controller_test.rb
@@ -1,0 +1,10 @@
+require "test_helper"
+
+class Admin::DashboardControllerTest < ActionController::TestCase
+
+  test "GET index renders the view" do
+    get :index
+    assert_response :ok
+    assert_template :index
+  end
+end
